test: split server tests
